requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
Pillow
pytesseract
opencv-python
numpy
pyperclip
pyautogui==0.9.54
pillow==10.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
flask==2.3.3
gunicorn==21.2.0